refactor(mcp/tools): extract the tool catalog (names/list/description/schema) to catalog.rs
tools/mod.rs was 1050 lines mixing request types, dispatch, and the tool registry.
Move TOOL_NAMES + list_tools + description() (130-line catalog) + schema() into
catalog.rs and re-export. mod.rs drops to 833 (request types + dispatch). No
behavior change; the catalog is now one cohesive, separately-evolving surface.
